image to np.float32 for imshow

+++ S9/utils.py
@@ -2,6 +2,7 @@ import torch
 import matplotlib.pyplot as plt
 from cuda import enable_cuda
 import cv2
+import numpy as np
 
 
 def set_seed(value=123):
@@ -86,7 +87,7 @@ def plot_results(data, classes=None):
             row_count += 1
         axs[row_count][idx % 5].axis('off')
         axs[row_count][idx % 5].set_title(f'Label: {label}\nPrediction: {prediction}')
-        axs[row_count][idx % 5].imshow(cv2.cvtColor(result['image'][0], cv2.COLOR_BGR2RGB))
+        axs[row_count][idx % 5].imshow(cv2.cvtColor(np.float32(result['image'][0]), cv2.COLOR_BGR2RGB))
 
     plt.show()
     fig.savefig(f'incorrect_predictions.png', bbox_inches='tight')
